feat(audit-log): restructure audit logging services and introduce new interfaces
- Created AuditLogModule to encapsulate audit logging functionality.
- Introduced IAuditLogEntry and related interfaces for better type safety.
- Implemented AuditLoggerService for logging events with methods for various actions.
- Added AuditQueryService for searching and retrieving logs with filters.
- Developed AuditReportingService for generating reports and statistics from logs.
- Implemented AuditExportService for exporting logs in JSON and CSV formats.
- Added unit tests for AuditLoggerService to ensure logging functionality works as expected.
- Updated imports in sensitive-operations.service.ts and masking-audit.service.ts to use new interfaces.
