+++ src/audit-log/audit-log.service.ts
@@ -1,116 +1,47 @@
-import { Injectable, Logger } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, Between, MoreThan } from 'typeorm';
+import { Injectable } from '@nestjs/common';
+import { AuditAction, AuditSeverity } from './enums/audit-action.enum';
 import { AuditLog } from './audit-log.entity';
-import { AuditAction, AuditSeverity, AuditCategory } from './enums/audit-action.enum';
-import { ConfigService } from '@nestjs/config';
-import { sanitizePii } from '../common/utils/pii-sanitizer.utils';
+import { AuditLoggerService } from './services/audit-logger.service';
+import { AuditQueryService } from './services/audit-query.service';
+import { AuditReportingService } from './services/audit-reporting.service';
+import { AuditExportService } from './services/audit-export.service';
 import {
-  buildRetentionCutoff,
-  buildRetentionUntil,
-  resolveRetentionDays,
-} from '../middleware/audit/log-retention.policy';
+  IAuditLogEntry,
+  IAuditLogSearchFilters,
+  IAuditLogSearchResult,
+  IAuditReport,
+} from './interfaces/audit-log.interfaces';
 
-export interface IAuditLogEntry {
-  userId?: string;
-  userEmail?: string;
-  action: AuditAction;
-  category: AuditCategory;
-  severity?: AuditSeverity;
-  entityType?: string;
-  entityId?: string;
-  description?: string;
-  metadata?: Record<string, unknown>;
-  oldValues?: Record<string, unknown>;
-  newValues?: Record<string, unknown>;
-  ipAddress?: string;
-  userAgent?: string;
-  sessionId?: string;
-  requestId?: string;
-  apiEndpoint?: string;
-  httpMethod?: string;
-  statusCode?: number;
-  responseTimeMs?: number;
-  tenantId?: string;
-}
-
-export interface IAuditLogSearchFilters {
-  userId?: string;
-  userEmail?: string;
-  actions?: AuditAction[];
-  categories?: AuditCategory[];
-  severities?: AuditSeverity[];
-  entityType?: string;
-  entityId?: string;
-  ipAddress?: string;
-  sessionId?: string;
-  tenantId?: string;
-  startDate?: Date;
-  endDate?: Date;
-  apiEndpoint?: string;
-  httpMethod?: string;
-  statusCode?: number;
-}
-
-export interface IAuditLogSearchResult {
-  logs: AuditLog[];
-  total: number;
-  page: number;
-  limit: number;
-  totalPages: number;
-}
-
-export interface IAuditReport {
-  period: { start: Date; end: Date };
-  totalEvents: number;
-  eventsByCategory: Record<string, number>;
-  eventsByAction: Record<string, number>;
-  eventsBySeverity: Record<string, number>;
-  topUsers: Array<{ userId: string; userEmail: string; count: number }>;
-  topEndpoints: Array<{ endpoint: string; count: number }>;
-  failedActions: Array<{ action: string; count: number }>;
-}
+// Re-export interfaces for backward compatibility
+export {
+  IAuditLogEntry,
+  IAuditLogSearchFilters,
+  IAuditLogSearchResult,
+  IAuditReport,
+} from './interfaces/audit-log.interfaces';
 
 /**
- * Provides audit Log operations.
+ * Provides audit logging operations.
+ * Acts as a facade delegating to specialized services following Single Responsibility Principle.
+ * - AuditLoggerService: Handles logging operations
+ * - AuditQueryService: Handles search and query operations
+ * - AuditReportingService: Handles report generation and statistics
+ * - AuditExportService: Handles export to various formats
  */
 @Injectable()
 export class AuditLogService {
-  private readonly logger = new Logger(AuditLogService.name);
-  private readonly retentionDays: number;
   constructor(
-    @InjectRepository(AuditLog)
-    private readonly auditRepo: Repository<AuditLog>,
-    private readonly configService: ConfigService,
-  ) {
-    this.retentionDays = resolveRetentionDays(
-      this.configService.get<number>('AUDIT_LOG_RETENTION_DAYS', 365),
-    );
-  }
+    private readonly loggerService: AuditLoggerService,
+    private readonly queryService: AuditQueryService,
+    private readonly reportingService: AuditReportingService,
+    private readonly exportService: AuditExportService,
+  ) {}
 
   /**
    * Log an audit event
    */
   async log(entry: IAuditLogEntry): Promise<AuditLog> {
-    const retentionUntil = buildRetentionUntil(this.retentionDays);
-
-    const log = this.auditRepo.create({
-      ...entry,
-      severity: entry.severity || AuditSeverity.INFO,
-      retentionUntil,
-    });
-
-    try {
-      const saved = await this.auditRepo.save(log);
-      this.logger.debug(
-        `Audit log created: ${log.action} - ${sanitizePii(log.description || 'no description')}`,
-      );
-      return saved;
-    } catch (error) {
-      this.logger.error('Failed to create audit log:', error);
-      // Don't throw - audit logging should not break main functionality
-      return log;
-    }
+    return this.loggerService.log(entry);
   }
 
   /**
@@ -125,16 +56,7 @@ export class AuditLogService {
     metadata?: Record<string, unknown>,
     severity: AuditSeverity = AuditSeverity.INFO,
   ): Promise<AuditLog> {
-    return this.log({
-      userId: userId || undefined,
-      userEmail: userEmail || undefined,
-      action,
-      category: AuditCategory.AUTHENTICATION,
-      severity,
-      ipAddress,
-      userAgent,
-      metadata,
-    });
+    return this.loggerService.logAuth(action, userId, userEmail, ipAddress, userAgent, metadata, severity);
   }
 
   /**
@@ -151,19 +73,17 @@ export class AuditLogService {
     ipAddress?: string,
     description?: string,
   ): Promise<AuditLog> {
-    return this.log({
+    return this.loggerService.logDataChange(
+      action,
       userId,
       userEmail,
-      action,
-      category: AuditCategory.DATA_MODIFICATION,
-      severity: AuditSeverity.INFO,
       entityType,
       entityId,
-      description,
       oldValues,
       newValues,
       ipAddress,
-    });
+      description,
+    );
   }
 
   /**
@@ -180,19 +100,9 @@ export class AuditLogService {
     userAgent: string,
     requestId?: string,
   ): Promise<AuditLog> {
-    const severity =
-      statusCode >= 500
-        ? AuditSeverity.ERROR
-        : statusCode >= 400
-          ? AuditSeverity.WARNING
-          : AuditSeverity.INFO;
-
-    return this.log({
-      userId: userId || undefined,
-      userEmail: userEmail || undefined,
-      action: AuditAction.API_CALLED,
-      category: AuditCategory.DATA_ACCESS,
-      severity,
+    return this.loggerService.logApiAccess(
+      userId,
+      userEmail,
       apiEndpoint,
       httpMethod,
       statusCode,
@@ -200,7 +110,7 @@ export class AuditLogService {
       ipAddress,
       userAgent,
       requestId,
-    });
+    );
   }
 
   /**
@@ -215,17 +125,7 @@ export class AuditLogService {
     description: string,
     metadata?: Record<string, unknown>,
   ): Promise<AuditLog> {
-    return this.log({
-      userId: userId || undefined,
-      userEmail: userEmail || undefined,
-      action,
-      category: AuditCategory.SECURITY,
-      severity: AuditSeverity.WARNING,
-      ipAddress,
-      userAgent,
-      description,
-      metadata,
-    });
+    return this.loggerService.logSecurityEvent(action, userId, userEmail, ipAddress, userAgent, description, metadata);
   }
 
   /**
@@ -236,112 +136,28 @@ export class AuditLogService {
     page: number = 1,
     limit: number = 50,
   ): Promise<IAuditLogSearchResult> {
-    const queryBuilder = this.auditRepo.createQueryBuilder('audit');
-
-    // Apply filters
-    if (filters.userId) {
-      queryBuilder.andWhere('audit.userId = :userId', { userId: filters.userId });
-    }
-
-    if (filters.userEmail) {
-      queryBuilder.andWhere('audit.userEmail = :userEmail', { userEmail: filters.userEmail });
-    }
-    if (filters.actions && filters.actions.length > 0) {
-      queryBuilder.andWhere('audit.action IN (:...actions)', { actions: filters.actions });
-    }
-    if (filters.categories && filters.categories.length > 0) {
-      queryBuilder.andWhere('audit.category IN (:...categories)', {
-        categories: filters.categories,
-      });
-    }
-    if (filters.severities && filters.severities.length > 0) {
-      queryBuilder.andWhere('audit.severity IN (:...severities)', {
-        severities: filters.severities,
-      });
-    }
-    if (filters.entityType) {
-      queryBuilder.andWhere('audit.entityType = :entityType', { entityType: filters.entityType });
-    }
-    if (filters.entityId) {
-      queryBuilder.andWhere('audit.entityId = :entityId', { entityId: filters.entityId });
-    }
-    if (filters.ipAddress) {
-      queryBuilder.andWhere('audit.ipAddress = :ipAddress', { ipAddress: filters.ipAddress });
-    }
-    if (filters.sessionId) {
-      queryBuilder.andWhere('audit.sessionId = :sessionId', { sessionId: filters.sessionId });
-    }
-    if (filters.tenantId) {
-      queryBuilder.andWhere('audit.tenantId = :tenantId', { tenantId: filters.tenantId });
-    }
-    if (filters.apiEndpoint) {
-      queryBuilder.andWhere('audit.apiEndpoint LIKE :apiEndpoint', {
-        apiEndpoint: `%${filters.apiEndpoint}%`,
-      });
-    }
-    if (filters.httpMethod) {
-      queryBuilder.andWhere('audit.httpMethod = :httpMethod', { httpMethod: filters.httpMethod });
-    }
-    if (filters.statusCode) {
-      queryBuilder.andWhere('audit.statusCode = :statusCode', { statusCode: filters.statusCode });
-    }
-    if (filters.startDate && filters.endDate) {
-      queryBuilder.andWhere('audit.timestamp BETWEEN :startDate AND :endDate', {
-        startDate: filters.startDate,
-        endDate: filters.endDate,
-      });
-    } else if (filters.startDate) {
-      queryBuilder.andWhere('audit.timestamp >= :startDate', { startDate: filters.startDate });
-    } else if (filters.endDate) {
-      queryBuilder.andWhere('audit.timestamp <= :endDate', { endDate: filters.endDate });
-    }
-
-    queryBuilder.orderBy('audit.timestamp', 'DESC');
-
-    const total = await queryBuilder.getCount();
-    const skip = (page - 1) * limit;
-    queryBuilder.skip(skip).take(limit);
-    const logs = await queryBuilder.getMany();
-
-    return {
-      logs,
-      total,
-      page,
-      limit,
-      totalPages: Math.ceil(total / limit),
-    };
+    return this.queryService.search(filters, page, limit);
   }
 
   /**
    * Find all logs (with limit)
    */
   async findAll(limit: number = 100): Promise<AuditLog[]> {
-    return this.auditRepo.find({
-      order: { timestamp: 'DESC' },
-      take: limit,
-    });
+    return this.queryService.findAll(limit);
   }
 
   /**
    * Find logs by user
    */
   async findByUser(userId: string, limit: number = 100): Promise<AuditLog[]> {
-    return this.auditRepo.find({
-      where: { userId },
-      order: { timestamp: 'DESC' },
-      take: limit,
-    });
+    return this.queryService.findByUser(userId, limit);
   }
 
   /**
    * Find logs by action
    */
   async findByAction(action: AuditAction, limit: number = 100): Promise<AuditLog[]> {
-    return this.auditRepo.find({
-      where: { action },
-      order: { timestamp: 'DESC' },
-      take: limit,
-    });
+    return this.queryService.findByAction(action, limit);
   }
 
   /**
@@ -352,230 +168,49 @@ export class AuditLogService {
     entityId: string,
     limit: number = 100,
   ): Promise<AuditLog[]> {
-    return this.auditRepo.find({
-      where: { entityType, entityId },
-      order: { timestamp: 'DESC' },
-      take: limit,
-    });
+    return this.queryService.findByEntity(entityType, entityId, limit);
   }
 
   /**
    * Find logs by IP address
    */
   async findByIpAddress(ipAddress: string, limit: number = 100): Promise<AuditLog[]> {
-    return this.auditRepo.find({
-      where: { ipAddress },
-      order: { timestamp: 'DESC' },
-      take: limit,
-    });
+    return this.queryService.findByIpAddress(ipAddress, limit);
   }
 
   /**
    * Find logs by date range
    */
   async findByDateRange(startDate: Date, endDate: Date, limit: number = 1000): Promise<AuditLog[]> {
-    return this.auditRepo.find({
-      where: {
-        timestamp: Between(startDate, endDate),
-      },
-      order: { timestamp: 'DESC' },
-      take: limit,
-    });
+    return this.queryService.findByDateRange(startDate, endDate, limit);
   }
 
   /**
    * Generate audit report
    */
   async generateReport(startDate: Date, endDate: Date): Promise<IAuditReport> {
-    const queryBuilder = this.auditRepo.createQueryBuilder('audit');
-    queryBuilder.where('audit.timestamp BETWEEN :startDate AND :endDate', {
-      startDate,
-      endDate,
-    });
-
-    const totalEvents = await queryBuilder.getCount();
-
-    // Events by category
-    const categoryStats = await this.auditRepo
-      .createQueryBuilder('audit')
-      .select('audit.category', 'category')
-      .addSelect('COUNT(*)', 'count')
-      .where('audit.timestamp BETWEEN :startDate AND :endDate', { startDate, endDate })
-      .groupBy('audit.category')
-      .getRawMany();
-
-    const eventsByCategory: Record<string, number> = {};
-    categoryStats.forEach((stat) => {
-      eventsByCategory[stat.category] = parseInt(stat.count, 10);
-    });
-
-    // Events by action
-    const actionStats = await this.auditRepo
-      .createQueryBuilder('audit')
-      .select('audit.action', 'action')
-      .addSelect('COUNT(*)', 'count')
-      .where('audit.timestamp BETWEEN :startDate AND :endDate', { startDate, endDate })
-      .groupBy('audit.action')
-      .getRawMany();
-
-    const eventsByAction: Record<string, number> = {};
-    actionStats.forEach((stat) => {
-      eventsByAction[stat.action] = parseInt(stat.count, 10);
-    });
-
-    // Events by severity
-    const severityStats = await this.auditRepo
-      .createQueryBuilder('audit')
-      .select('audit.severity', 'severity')
-      .addSelect('COUNT(*)', 'count')
-      .where('audit.timestamp BETWEEN :startDate AND :endDate', { startDate, endDate })
-      .groupBy('audit.severity')
-      .getRawMany();
-
-    const eventsBySeverity: Record<string, number> = {};
-    severityStats.forEach((stat) => {
-      eventsBySeverity[stat.severity] = parseInt(stat.count, 10);
-    });
-
-    // Top users
-    const topUsers = await this.auditRepo
-      .createQueryBuilder('audit')
-      .select('audit.userId', 'userId')
-      .addSelect('audit.userEmail', 'userEmail')
-      .addSelect('COUNT(*)', 'count')
-      .where('audit.timestamp BETWEEN :startDate AND :endDate', { startDate, endDate })
-      .andWhere('audit.userId IS NOT NULL')
-      .groupBy('audit.userId')
-      .addGroupBy('audit.userEmail')
-      .orderBy('count', 'DESC')
-      .limit(10)
-      .getRawMany();
-
-    // Top endpoints
-    const topEndpoints = await this.auditRepo
-      .createQueryBuilder('audit')
-      .select('audit.apiEndpoint', 'endpoint')
-      .addSelect('COUNT(*)', 'count')
-      .where('audit.timestamp BETWEEN :startDate AND :endDate', { startDate, endDate })
-      .andWhere('audit.apiEndpoint IS NOT NULL')
-      .groupBy('audit.apiEndpoint')
-      .orderBy('count', 'DESC')
-      .limit(10)
-      .getRawMany();
-
-    // Failed actions (status code >= 400)
-    const failedActions = await this.auditRepo
-      .createQueryBuilder('audit')
-      .select('audit.action', 'action')
-      .addSelect('COUNT(*)', 'count')
-      .where('audit.timestamp BETWEEN :startDate AND :endDate', { startDate, endDate })
-      .andWhere('audit.statusCode >= 400')
-      .groupBy('audit.action')
-      .orderBy('count', 'DESC')
-      .limit(10)
-      .getRawMany();
-
-    return {
-      period: { start: startDate, end: endDate },
-      totalEvents,
-      eventsByCategory,
-      eventsByAction,
-      eventsBySeverity,
-      topUsers: topUsers.map((u) => ({
-        userId: u.userId,
-        userEmail: u.userEmail || 'Unknown',
-        count: parseInt(u.count, 10),
-      })),
-      topEndpoints: topEndpoints.map((e) => ({
-        endpoint: e.endpoint,
-        count: parseInt(e.count, 10),
-      })),
-      failedActions: failedActions.map((f) => ({
-        action: f.action,
-        count: parseInt(f.count, 10),
-      })),
-    };
+    return this.reportingService.generateReport(startDate, endDate);
   }
 
   /**
    * Apply retention policy - delete old logs
    */
   async applyRetentionPolicy(): Promise<number> {
-    const cutoffDate = buildRetentionCutoff(this.retentionDays);
-
-    const result = await this.auditRepo
-      .createQueryBuilder()
-      .delete()
-      .where('timestamp < :cutoffDate', { cutoffDate })
-      .orWhere('retentionUntil < NOW()')
-      .execute();
-
-    const deletedCount = result.affected || 0;
-    this.logger.log(`Applied retention policy: deleted ${deletedCount} old audit logs`);
-    return deletedCount;
+    return this.loggerService.applyRetentionPolicy();
   }
 
   /**
    * Export logs to JSON
    */
   async exportToJson(filters: IAuditLogSearchFilters): Promise<string> {
-    const { logs } = await this.search(filters, 1, 10000);
-    return JSON.stringify(logs, null, 2);
+    return this.exportService.exportToJson(filters);
   }
 
   /**
    * Export logs to CSV
    */
   async exportToCsv(filters: IAuditLogSearchFilters): Promise<string> {
-    const { logs } = await this.search(filters, 1, 10000);
-
-    const headers = [
-      'timestamp',
-      'userId',
-      'userEmail',
-      'action',
-      'category',
-      'severity',
-      'entityType',
-      'entityId',
-      'description',
-      'ipAddress',
-      'userAgent',
-      'apiEndpoint',
-      'httpMethod',
-      'statusCode',
-    ];
-
-    const rows = logs.map((log) => [
-      log.timestamp.toISOString(),
-      log.userId || '',
-      log.userEmail || '',
-      log.action,
-      log.category,
-      log.severity,
-      log.entityType || '',
-      log.entityId || '',
-      log.description || '',
-      log.ipAddress || '',
-      log.userAgent || '',
-      log.apiEndpoint || '',
-      log.httpMethod || '',
-      log.statusCode || '',
-    ]);
-
-    const escapeCsv = (value: string | number) => {
-      const str = String(value);
-      if (str.includes(',') || str.includes('"') || str.includes('\n')) {
-        return `"${str.replace(/"/g, '""')}"`;
-      }
-      return str;
-    };
-
-    const csvContent = [headers.join(','), ...rows.map((row) => row.map(escapeCsv).join(','))].join(
-      '\n',
-    );
-
-    return csvContent;
+    return this.exportService.exportToCsv(filters);
   }
 
   /**
@@ -589,32 +224,6 @@ export class AuditLogService {
     criticalEvents: number;
     errorEvents: number;
   }> {
-    const now = new Date();
-    const today = new Date(now.getFullYear(), now.getMonth(), now.getDate());
-    const weekAgo = new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000);
-    const monthAgo = new Date(now.getTime() - 30 * 24 * 60 * 60 * 1000);
-
-    const [totalLogs, logsToday, logsThisWeek, logsThisMonth, criticalEvents, errorEvents] =
-      await Promise.all([
-        this.auditRepo.count(),
-        this.auditRepo.count({ where: { timestamp: MoreThanOrEqual(today) } }),
-        this.auditRepo.count({ where: { timestamp: MoreThanOrEqual(weekAgo) } }),
-        this.auditRepo.count({ where: { timestamp: MoreThanOrEqual(monthAgo) } }),
-        this.auditRepo.count({ where: { severity: AuditSeverity.CRITICAL } }),
-        this.auditRepo.count({ where: { severity: AuditSeverity.ERROR } }),
-      ]);
-
-    return {
-      totalLogs,
-      logsToday,
-      logsThisWeek,
-      logsThisMonth,
-      criticalEvents,
-      errorEvents,
-    };
+    return this.reportingService.getStatistics();
   }
-}
-// Helper function for date comparison
-function MoreThanOrEqual(date: Date) {
-  return MoreThan(date);
-}
+
